Issues: Fixes #6
Problem:
The issue here is that we're not generating .deb and .rpm files for
release.

Analysis:
This commit contains artifacts, scripts, and variable additions to the
.travis.yml file that will help to orchestrate the creation of the .rpm
and .deb files.  This largely uses the same functionality generated to
assist in the
[f5-common-python](https://github.com/F5Networks/f5-common-python)
library.

Tests:
This is, in and of itself, a test.